+++ reana_client/cli/workflow.py
@@ -244,7 +244,6 @@ def workflows_list(  # noqa: C901
 
         data = []
         for workflow in response:
-            workflow["size"] = workflow["size"][human_readable_or_raw]
             name, run_number = get_workflow_name_and_run_number(workflow["name"])
             workflow["name"] = name
             workflow["run_number"] = run_number
@@ -271,19 +270,30 @@ def workflows_list(  # noqa: C901
             data.append(row)
 
         # Sort by given column, making sure that `None` is at the bottom of the list.
+        def get_sort_key(x, column_id):
+            if sort_columm_name == "run_number":
+                return list(map(int, x[column_id].split(".")))
+            elif sort_columm_name in workspace_size_header:
+                return x[column_id]["raw"]
+            return x[column_id] is not None, x[column_id]
+
         sort_column_id = 2
         if sort_columm_name.lower() in headers[type]:
             sort_column_id = headers[type].index(sort_columm_name.lower())
         data = sorted(
             data,
-            key=lambda x: (x[sort_column_id] is not None, x[sort_column_id]),
+            key=lambda x: get_sort_key(x, sort_column_id),
             reverse=True,
         )
 
-        # Substitute `None` with "-"
         for row in data:
             for i, value in enumerate(row):
-                row[i] = value or "-"
+                # Substitute `None` with "-"
+                content = value or "-"
+                # Replace raw size with human-readable size if requested
+                if headers[type][i] in workspace_size_header:
+                    content = value.get(human_readable_or_raw)
+                row[i] = content
 
         workflow_ids = ["{0}.{1}".format(w[0], w[1]) for w in data]
         if os.getenv("REANA_WORKON", "") in workflow_ids: