+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes
 
 Version 0.9.1 (UNRELEASED)
 --------------------------
+- Fixes ``list`` command to correctly list workflows when sorting them by their run number or by the size of their workspace.
 - Fixes ``du`` command help message typo.
 
 Version 0.9.0 (2023-01-26)